requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
SQLAlchemy
psycopg2-binary
python-multipart==0.0.6
PyMuPDF==1.22.5       # fitz for PDF extraction
requests==2.31.0
pydantic==1.10.11
python-docx==0.8.11